In-progress draft of installation logistics modeling framework
This is focused on a framework for representing installation actions
to start with. Located in Irma folder for now.
Irma.py has the basic script for trying things out, and more
importantly a draft Scenario class that manages actions and related
info. Includes Rudy's graph theory representation and plotting function.
Actions.py has the class that covers individual actions.
Four YAML files give an outline/beginning of a framework to
represent how various vessels/assets, capabilities, and actions
can be interrelated and organized.